authenrize gate and proxi

+++ laravel_QA/app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,12 @@ use App\Http\Requests\AskQquestionRequest;
 
 class QuestionsController extends Controller
 {
+    public function __construct()
+    {   
+        // k cho thêm câu hỏi, nhưng vẫn cho xem
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -67,6 +73,11 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        // if(\Gate::denies('update-question', $question)) {
+        //     abort(403, 'hihi');
+        // }
+
+        $this->authorize('update', $question);
         return view('questions.edit', compact('question'));
     }
 
@@ -79,6 +90,11 @@ class QuestionsController extends Controller
      */
     public function update(Request $request, Question $question)
     {
+
+        // if(\Gate::denies('update-question', $question)) {
+        //     abort(403, 'hihi');
+        // }
+        $this->authorize('update', $question);
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success' , 'Your question has been updated.');
@@ -92,6 +108,11 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        // if(\Gate::denies('delete-question', $question)) {
+        //     abort(403, 'ahhi');
+        // }
+
+        $this->authorize('delete', $question);
         $question->delete();
 
         return redirect('/questions')->with('success', 'Your question has delete.');
